+++ assets/helpers/theGrid.js
@@ -22,12 +22,11 @@ export const imageCatalogue: {
   protestorsNarrow: 'd1a7088f8f2a367b0321528f081777c9b5618412/0_0_3578_2013',
   premiumTierCircle: '3e3b59228c3467b01dd09b9f79de28c586fe0ea6/0_0_825_825',
   dailyEditionCircle: '64ba1a800e4c2975054b33ba07c4fe3c64d4b5f4/0_0_825_825',
-  digitalCircle: '639c3abc4c09281aedb515d684ac4053ef38a1df/0_0_825_825',
-  digitalCircleAlt: '0b5ded2325275e25f5b38198766c57363bddfa21/0_0_825_825',
-  digitalCircleInternational: '7d404c1920f065c1b7e71b903cc3899f388acb22/0_0_825_825',
+  digitalCircleOrange: '0b5ded2325275e25f5b38198766c57363bddfa21/0_0_825_825',
+  digitalCirclePink: '7d404c1920f065c1b7e71b903cc3899f388acb22/0_0_825_825',
   paperCircle: 'c462d60f2962b745b1e206d5ede998dfb166a8ed/0_0_825_825',
-  paperDigitalCircle: 'd94c0f9bade09487b9afca5ee8149efb33f34ccf/0_0_825_825',
-  paperDigitalCircleAlt: '69d90e5d6fca261a227e47b311f80807b123c87b/0_0_825_825',
+  paperDigitalCircleOrange: 'd94c0f9bade09487b9afca5ee8149efb33f34ccf/0_0_825_825',
+  paperDigitalCirclePink: '69d90e5d6fca261a227e47b311f80807b123c87b/0_0_825_825',
   weeklyCircle: '13cfdcba3f0738b2ed54cad21d43ca87b4cb7855/0_0_825_825',
   premiumTier: 'fb0c788ddee28f8e0e66d814595cf81d6aa21ec6/0_0_644_448',
   premiumTierAU: '4b6fd9805c7d0a88b4b71a683c4b46279a410b9d/0_0_1610_1120',
